# phase-4/specs/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo API - Phase II
  description: REST API for multi-user todo application with JWT authentication
  version: 1.0.0
  contact:
    name: Evolution of Todo
    url: https://github.com/example/evolution-of-todo

servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server

tags:
  - name: Authentication
    description: User registration and login endpoints
  - name: Todos
    description: Todo CRUD operations

paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Creates a new user account and returns authentication token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
            example:
              email: user@example.com
              password: securepassword123
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticates user and returns JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
            example:
              email: user@example.com
              password: securepassword123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Returns the currently authenticated user's information
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /todos:
    get:
      tags:
        - Todos
      summary: List all todos
      description: Returns all todo tasks for the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TodoResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Todos
      summary: Create a todo
      description: Creates a new todo task for the authenticated user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoCreateRequest'
            example:
              title: Buy groceries
              description: Milk, eggs, bread
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /todos/{todoId}:
    get:
      tags:
        - Todos
      summary: Get a todo
      description: Returns a specific todo task by ID
      security:
        - BearerAuth: []
      parameters:
        - name: todoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the todo task
      responses:
        '200':
          description: Todo details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Todos
      summary: Update a todo
      description: Replaces an existing todo task with new data
      security:
        - BearerAuth: []
      parameters:
        - name: todoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the todo task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoUpdateRequest'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags:
        - Todos
      summary: Partial update a todo
      description: Updates specific fields of an existing todo task
      security:
        - BearerAuth: []
      parameters:
        - name: todoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the todo task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoPartialUpdateRequest'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Todos
      summary: Delete a todo
      description: Deletes an existing todo task
      security:
        - BearerAuth: []
      parameters:
        - name: todoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the todo task
      responses:
        '204':
          description: Todo deleted successfully
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    # Authentication Schemas
    UserRegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          example: securepassword123

    UserLoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: securepassword123

    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: user@example.com
        created_at:
          type: string
          format: date-time
          example: 2026-01-06T00:00:00Z

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: '#/components/schemas/UserResponse'

    # Todo Schemas
    TodoCreateRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          example: Buy groceries
        description:
          type: string
          nullable: true
          example: Milk, eggs, bread

    TodoUpdateRequest:
      type: object
      required:
        - title
        - description
        - completed
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          example: Updated title
        description:
          type: string
          nullable: true
          example: Updated description
        completed:
          type: boolean
          example: true

    TodoPartialUpdateRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          nullable: true
        description:
          type: string
          nullable: true
        completed:
          type: boolean
          nullable: true

    TodoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        user_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440001
        title:
          type: string
          example: Buy groceries
        description:
          type: string
          nullable: true
          example: Milk, eggs, bread
        completed:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: 2026-01-06T00:00:00Z
        updated_at:
          type: string
          format: date-time
          example: 2026-01-06T00:00:00Z

    # Error Schemas
    ErrorResponse:
      type: object
      properties:
        detail:
          oneOf:
            - type: string
              example: Not authenticated
            - type: array
              items:
                type: object
                properties:
                  loc:
                    type: array
                    items:
                      type: string
                  msg:
                    type: string
                  type:
                    type: string